back-end/requirements.txt
Flask==3.1.2
pytest==8.4.2
PyJWT==2.9.0
flask-cors==5.0.0